#13 Remove FirebaseUserDetails class
