+++ src/main/java/com/takefour/themoment/themoment/config/method/support/UserHandlerMethodArgumentResolver.java
@@ -4,13 +4,13 @@ import org.springframework.core.MethodParameter;
 import org.springframework.security.authentication.AuthenticationCredentialsNotFoundException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.User;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
 import com.takefour.themoment.themoment.config.method.annotation.CurrentUser;
+import com.takefour.themoment.themoment.model.User;
 
 
 public class UserHandlerMethodArgumentResolver implements HandlerMethodArgumentResolver {
