--- src/main/java/com/takefour/themoment/themoment/security/FirebaseUserDetails.java
@@ -1,56 +1,0 @@
-package com.takefour.themoment.themoment.security;
-
-import java.util.Collection;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.AuthorityUtils;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import com.takefour.themoment.themoment.model.User;
-
-/**
- * Created by hanbyeol on 2018. 1. 11..
- */
-public class FirebaseUserDetails implements UserDetails {
-
-	private User user;
-
-	public FirebaseUserDetails(User user) {
-		this.user = user;
-	}
-
-	@Override
-	public Collection<? extends GrantedAuthority> getAuthorities() {
-		return AuthorityUtils.createAuthorityList(user.getRole(), "USER");
-	}
-
-	@Override
-	public String getPassword() {
-		return user.getPassword();
-	}
-
-	@Override
-	public String getUsername() {
-		return user.getEmail();
-	}
-
-	@Override
-	public boolean isAccountNonExpired() {
-		return true;
-	}
-
-	@Override
-	public boolean isAccountNonLocked() {
-		return true;
-	}
-
-	@Override
-	public boolean isCredentialsNonExpired() {
-		return true;
-	}
-
-	@Override
-	public boolean isEnabled() {
-		return true;
-	}
-}
